Pull in updated verification server docker image
Pulls in an updated docker image for the verification server to
support the following additional endpoints:

1. /version/v1/registrationToken
2. /version/v1/tan
3. /version/v1/testresult

Each of them accepts a finite set of valid inputs as specified in the
cwa-verification-fake code.

This allows a client to complete all steps in the verification process
for either TeleTANs or GUIDs, all the way through to receiving a test
result.